add an option to always recheck process names
on each scrape.
Although this is less efficient, it is useful/needed if some processes use exec and
would otherwise be completey ignored when they still have a name that is not supposed
to be tracked when checked for the first time.
E.g. when starting some processes with a wrapper bash script that then `exec`s the actual
process later.